Time: 667 ms (97.76%) | Memory: 56.1 MB (67.73%) - LeetSync

+++ 1803-average-waiting-time/average-waiting-time.py
@@ -8,4 +8,4 @@ class Solution(object):
             else:
                 cur += vec[1]
                 sum += cur - vec[0]
-        return sum/len(customers)
+        return float(sum/len(customers))
